split out methods
