+++ index.js
@@ -1,7 +1,9 @@
 'use strict';
 
 var extend = require('extend-shallow');
-var snapdragon = require('snapdragon');
+var renderer = require('./lib/renderer');
+var parser = require('./lib/parser');
+var utils = require('./lib/utils');
 var makeReCache = {};
 var regexCache = {};
 
@@ -47,527 +49,14 @@ function extglob(pattern, options) {
 }
 
 /**
- * Parse an extglob pattern into an AST that can be passed
- * to the [.render](#render) method.
- *
- * ```js
- * var extglob = require('extglob');
- * var ast = extglob.parse('!(foo|bar)');
- * ```
- * @param {String} `str` Extglob pattern
- * @param {Object} `options`
- * @return {Object} AST
- * @api public
+ * Expose parser and render methods
  */
 
-extglob.parse = function(pattern, options) {
-  var opts = extend({}, options);
-  var parser = new snapdragon.Parser(opts);
-  var sets = {bracket: [], brace: [], paren: []};
-
-  var first = pattern.charAt(0);
-  if (first !== '.') {
-    if (!/^([!^@*?+]\()/.test(pattern)) {
-      parser.prefix = '(?!\\.)(?=.)';
-    } else {
-      parser.prefix = '';
-    }
-  }
-
-  parser
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\^(?=.)/);
-      if (!m) return;
-      return pos({
-        type: 'boundary.start',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/(?=.\$)\$$/);
-      if (!m) return;
-      return pos({
-        type: 'boundary.end',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[!](?!\()/);
-      if (!m) return;
-
-      if (this.nodes.length === 0) {
-        this.isNegated = true;
-      }
-
-      return pos({
-        type: 'exclamation',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\\(.)/);
-      if (!m) return;
-      return pos({
-        type: 'escaped',
-        val: m[0],
-        esc: m[1]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[.]/);
-      if (!m) return;
-      return pos({
-        type: 'dot',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\[\]\]/);
-      if (!m) return;
-      return pos({
-        type: 'brackets.literal',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\[\](?!.*\])/);
-      if (!m) return;
-      return pos({
-        type: 'brackets.empty',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\[([!^]*)(.)/);
-      if (!m) return;
-      var token = {
-        type: 'brackets.open',
-        prefix: m[1],
-        infix: m[2],
-        sets: sets,
-        val: m[0]
-      };
-      this.sets.brackets.push(token);
-      return pos(token);
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\]/);
-      if (!m) return;
-
-      var token = this.sets.brackets.pop() || {};
-      var prefix = token.prefix;
-
-      if (typeof prefix === 'undefined') {
-        if (parser.options.strictbrackets === true) {
-          throw new Error('missing opening bracket');
-        }
-        token = {type: 'brackets.open', val: '['};
-      }
-
-      return pos({
-        type: 'brackets.close',
-        prev: token,
-        prefix: prefix,
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\(\)/);
-      if (!m) return;
-      var token = {
-        type: 'parens.empty',
-        prefix: m[1],
-        val: m[0]
-      };
-      this.sets.parens.push(token);
-      return pos(token);
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^([!^@*?+])\(/);
-      if (!m) return;
-      var token = {
-        type: 'extglob.open',
-        idx: this.nodes.length,
-        prefix: m[1],
-        sets: sets,
-        val: m[0]
-      };
-      this.sets.parens.push(token);
-      return pos(token);
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^(?![!^@*?+])\(/);
-      if (!m) return;
-      var token = {
-        type: 'parens.open',
-        val: m[0]
-      };
-      this.sets.parens.push(token);
-      return pos(token);
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\)/);
-      if (!m) return;
-
-      var token = this.sets.parens.pop() || {};
-      var prefix = token.prefix;
-      var val = '';
-
-      if (typeof prefix === 'undefined') {
-        if (parser.options.strictparens === true) {
-          throw new Error('missing opening paren');
-        }
-        token = {type: 'parens.open', val: '('};
-      }
-
-      switch (prefix) {
-        case '^':
-        case '!':
-          if (this.input) {
-            var next = this.input.charAt(0);
-            if (/\w/i.test(next)) {
-              val = ')).*?)';
-            } else if (next === '?') {
-              val = ')\\b)[^/])';
-            } else if (next === '*') {
-              val = '(?=.))$).*)';
-            } else {
-              if (next === this.input) {
-                var infix = this.nodes[token.idx + 1];
-                if (infix && /[*+]/.test(infix.val)) {
-                  this.strictclose = false;
-                }
-              }
-              val = ')\\b).*)';
-            }
-          } else {
-            val = ')$).*)';
-          }
-          break;
-        case '+':
-          val = ')+';
-          break;
-        case '*':
-          val = ')*';
-          break;
-        case '?':
-        case '@':
-        default: {
-          val = ')';
-          break;
-        }
-      }
-
-      return pos({
-        type: 'extglob.close',
-        prefix: prefix,
-        val: val
-      });
-    })
-
-    /**
-     * Match `**` followed by anything by `(`
-     */
-
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[*]{2,}(?!\()/);
-      if (!m) return;
-      return pos({
-        type: 'globstar',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[*][.][*]/);
-      if (!m) return;
-      return pos({
-        type: 'globname',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[*](?!\()/);
-      if (!m) return;
-      var prev = this.nodes[this.nodes.length - 1];
-      var val = m[0];
-      return pos({
-        type: 'star',
-        prev: prev,
-        val: val
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[+](?!\()/);
-      if (!m) return;
-      var prev = this.nodes[this.nodes.length - 1];
-      return pos({
-        type: 'plus',
-        prev: prev,
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^(?=.)[?][|][?](?=.)/);
-      if (!m) return;
-      return pos({
-        type: 'qmark.or',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^(.)?\?+(?![|])/);
-      if (!m) return;
-
-      return pos({
-        type: 'qmark',
-        prefix: m[1],
-        val: m[0]
-      });
-    })
-   .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[|]/);
-      if (!m) return;
-      return pos({
-        type: 'pipe',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[-]/);
-      if (!m) return;
-      return pos({
-        type: 'dash',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^[,]/);
-      if (!m) return;
-      return pos({
-        type: 'comma',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\d+/);
-      if (!m) return;
-      return pos({
-        type: 'number',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\w+/);
-      if (!m) return;
-      return pos({
-        type: 'text',
-        val: m[0]
-      });
-    })
-    .use(function() {
-      var pos = this.position();
-      var m = this.match(/^\//);
-      if (!m) return;
-      return pos({
-        type: 'slash',
-        val: m[0]
-      });
-    });
-
-  var ast = parser.parse(pattern);
-  if (ast.isNegated) {
-    ast.nodes.push({
-      type: 'end',
-      val: ')$).*'
-    });
-  }
-
-  fixMissingBrackets(parser);
-  return ast;
-};
-
-/**
- * Fix missing => `[]`, `{}`, `()`
- */
-
-function fixMissingBrackets(parser) {
-  for (var key in parser.sets) {
-    if (parser.sets.hasOwnProperty(key)) {
-      parser.sets[key].forEach(function(token) {
-        if (token.type === key + '.open') {
-          if (parser.options['strict' + key] === true) {
-            throw parser.error(['missing opening', key + ': "' + parser.parsed + '"'], token);
-          }
-          token.val = '\\' + token.val;
-          token.esc = true;
-        }
-      });
-    }
-  }
-}
-
-/**
- * Render a string from an extglob AST.
- *
- * ```js
- * var extglob = require('extglob');
- * var ast = extglob.parse('!(foo|bar)');
- * var str = extglob.render(ast);
- * ```
- * @param {Object} `ast` Extglob pattern
- * @param {Object} `options`
- * @return {String} Returns a regex-compatible string.
- * @api public
- */
-
-extglob.render = function(ast, options) {
-  if (ast == null || typeof ast !== 'object') {
-    throw new TypeError('expected ast to be an object');
-  }
-
-  function extglobOpen(node)  {
-    switch (node.prefix) {
-      case '!':
-      case '^':
-        return '(?:(?!(?:';
-      case '@':
-      case '+':
-      case '*':
-      case '?':
-        return '(?:';
-      default: {
-        return node.val;
-      }
-    }
-  }
-
-  var renderer = new snapdragon.Renderer(options)
-    .set('boundary.start', function(node)  {
-      return '\\b';
-    })
-    .set('boundary.end', function(node)  {
-      return '\\b';
-    })
-    .set('exclamation', function(node)  {
-      return '(?!^(?:';
-    })
-    .set('escaped', function(node)  {
-      return node.val;
-    })
-    .set('parens.open', extglobOpen)
-    .set('parens.empty', function(node)  {
-      return '.?';
-    })
-    .set('extglob.open', extglobOpen)
-    .set('extglob.close', function(node)  {
-      return node.val;
-    })
-    .set('brackets.empty', function(node)  {
-      return '\\[\\]';
-    })
-    .set('brackets.literal', function(node)  {
-      return '\\]';
-    })
-    .set('brackets.open', function(node)  {
-      if (/^[\[\]]/.test(node.infix)) {
-        node.infix = '\\' + node.infix;
-      }
-      return '[' + (node.prefix ? '^' : '') + node.infix;
-    })
-    .set('brackets.close', function(node)  {
-      return node.val;
-    })
-    .set('globstar', function(node)  {
-      return '(?:(?!(?:\\\/|^)\\.).)*?';
-    })
-    .set('globname', function(node)  {
-      return '([^.]*[.][^.]*)';
-    })
-    .set('star', function(node)  {
-      if (node.prev && node.prev.esc) {
-        return '*?';
-      } else {
-        return '[^/]*?';
-      }
-    })
-    .set('plus', function(node)  {
-      return node.val;
-    })
-    .set('qmark', function(node, nodes, idx)  {
-      var prev = nodes[idx - 1];
-      if (nodes.length === 1) {
-        return '[^/]{0,' + node.val.length + '}';
-      }
-      if (typeof prev === 'undefined' || prev.type === 'extglob.close') {
-        return node.prefix || '';
-      }
-      if (prev.type === 'escaped') {
-        return '[^/]';
-      }
-      if (prev.val === '*') {
-        return '[^/]';
-      }
-      return node.val;
-    })
-    .set('qmark.or', function(node)  {
-      return '?)|(?:';
-    })
-    .set('dash', function(node)  {
-      return '-';
-    })
-    .set('pipe', function(node)  {
-      return '|';
-    })
-    .set('comma', function(node)  {
-      return '|';
-    })
-    .set('dot', function(node, nodes, idx)  {
-      var prev = nodes[idx - 1] || {};
-      var val = prev.val || '';
-      if (idx === 0 || /\/$/.test(val) || this.options.dot === true) {
-        return '(?!(?:^|\\\/)\\.{1,2}(?:$|\\\/))';
-      }
-      return '(?=.)\\' + node.val;
-    })
-    .set('slash', function(node)  {
-      return '\\/';
-    })
-    .set('number', function(node)  {
-      return node.val;
-    })
-    .set('text', function(node)  {
-      return node.val;
-    })
-    .set('end', function(node)  {
-      return node.val;
-    });
-
-  return renderer.render(ast);
-};
+extglob.parser = parser;
+extglob.parse = parser.parse;
+extglob.renderer = renderer;
+extglob.stringify = renderer.stringify;
+extglob.render = renderer.render;
 
 /**
  * Create a regular expression from the given extglob `pattern`.
