+++ lib/utils.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var path = require('path');
+var util = require('util');
+var extend = require('extend-shallow');
+
+
+var utils = module.exports;
